Narrow MAX_SPLITS and MAX_MINTS from usize to u8
usize is not IDL-compatible, and these values are small enough
for u8. Adjust call sites to cast where needed.

+++ programs/flex/src/state/mod.rs
@@ -1,8 +1,8 @@
 use anchor_lang::prelude::*;
 
-pub const MAX_SPLITS: usize = 5;
+pub const MAX_SPLITS: u8 = 5;
 pub const MAX_PENDING_LIMIT: u16 = u16::MAX;
-pub const MAX_MINTS: usize = 8;
+pub const MAX_MINTS: u8 = 8;
 
 pub const MIN_DEADMAN_TIMEOUT_SLOTS: u64 = 1_000;
 pub const MAX_REFUND_TIMEOUT_SLOTS: u64 = 1_296_000;
@@ -61,6 +61,6 @@ pub struct PendingSettlement {
     pub submitted_at_slot: u64,
     pub session_key: Pubkey,
     pub split_count: u8,
-    pub splits: [SplitEntry; MAX_SPLITS],
+    pub splits: [SplitEntry; MAX_SPLITS as usize],
     pub bump: u8,
 }
